+++ spec/app/models/post.rb
@@ -17,6 +17,7 @@ class Post
 
   scope :recent, where(created_at: { "$lt" => Time.now, "$gt" => 30.days.ago })
   scope :posting, where(:content.in => [ "Posting" ])
+  scope :open, where(title: "open")
 
   validates_format_of :title, without: /\$\$\$/